// partner-search/frontend/src/app/api/model/address-dto.ts
/**
 * traefik-microstack API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {TelecommunicationDto} from './telecommunication-dto';
import {AddressTypeDto} from './address-type-dto';


/**
 * Address information containing location details and related telecommunications
 */
export interface AddressDto {
  /**
   * Unique address identifier - primary key. Optional for new addresses (auto-generated), required for updates.
   */
  addressId?: number;
  /**
   * Person ID - foreign key to ContactPersonEntity
   */
  personId?: number;
  /**
   * Address type identifier
   */
  addressTypeId?: number;
  /**
   * Postal code (required)
   */
  postalCode: string;
  /**
   * Post office box
   */
  poBox?: string;
  /**
   * Street name (required)
   */
  street: string;
  /**
   * House number
   */
  houseNumber: string;
  /**
   * City/Place name (required)
   */
  city: string;
  /**
   * Country identifier
   */
  countryId?: number;
  /**
   * Last modification date
   */
  modificationDate?: string;
  /**
   * ID of person who made the last modification
   */
  modifierId?: string;
  addressType?: AddressTypeDto;
  /**
   * List of telecommunication entries (phone, email, etc.) for this address
   */
  telecommunications?: Array<TelecommunicationDto>;
}